% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjusted_stats}
\alias{adjusted_stats}
\title{Adjusted Mean and Standard Deviation}
\usage{
adjusted_stats(mean_original, sd_original, a, b)
}
\arguments{
\item{mean_original}{The original mean of the dataset.}

\item{sd_original}{The original standard deviation of the dataset.}

\item{a}{The adjustment to be added to each value.}

\item{b}{The scaling factor to be applied after the adjustment.}
}
\value{
A list containing the adjusted mean and standard deviation.
}
\description{
Calculates the adjusted mean and standard deviation of a dataset after applying a linear transformation.
}
\examples{
adjusted_stats(mean_original = 5.5, sd_original = 1.8, a = 0.2, b = 1.4)
}
