% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyze_stem_leaf}
\alias{analyze_stem_leaf}
\title{Analyze Stem-and-Leaf Plot}
\usage{
analyze_stem_leaf(stems, leaves)
}
\arguments{
\item{stems}{A numeric vector of stem values.}

\item{leaves}{A list of numeric vectors representing leaves corresponding to each stem.}
}
\value{
A list containing the median and a description of the distribution's shape.
}
\description{
Calculates the median and describes the shape of the distribution from a stem-and-leaf plot.
}
\examples{
stems <- c(2, 3, 4, 5)
leaves <- list(
  c(3, 4, 6, 7, 7, 8, 8, 8, 9),
  c(2, 2, 3, 4, 6, 7, 8, 9),
  c(1, 2, 2, 3, 4),
  c(3, 5, 5, 6)
)
analyze_stem_leaf(stems, leaves)
}
