% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_between}
\alias{binomial_between}
\title{Binomial Probability: Between Two Values}
\usage{
binomial_between(n, p, k1, k2)
}
\arguments{
\item{n}{The number of trials.}

\item{p}{The probability of success in a single trial.}

\item{k1}{The lower bound for successes (inclusive).}

\item{k2}{The upper bound for successes (exclusive).}
}
\value{
The probability of successes between \( k1 \) and \( k2 \) (inclusive for \( k1 \), exclusive for \( k2 \)).
}
\description{
Calculates the probability of k1 <= X < k2 successes in a binomial distribution.
}
\examples{
binomial_between(n = 15, p = 0.7, k1 = 8, k2 = 11)
}
