% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_between_inclusive}
\alias{binomial_between_inclusive}
\title{Binomial Probability: Between Two Values (Inclusive)}
\usage{
binomial_between_inclusive(n, p, k1, k2)
}
\arguments{
\item{n}{The number of trials.}

\item{p}{The probability of success in a single trial.}

\item{k1}{The lower bound for successes (inclusive).}

\item{k2}{The upper bound for successes (inclusive).}
}
\value{
The probability of successes between \( k1 \) and \( k2 \), inclusive.
}
\description{
Calculates the probability of k1 <= X <= k2 successes in a binomial distribution.
}
\examples{
binomial_between_inclusive(n = 20, p = 0.2, k1 = 7, k2 = 8)
}
