% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_exactly}
\alias{binomial_exactly}
\title{Binomial Probability: Exactly}
\usage{
binomial_exactly(n, p, k)
}
\arguments{
\item{n}{The number of trials.}

\item{p}{The probability of success in a single trial.}

\item{k}{The number of successes for which the probability is calculated.}
}
\value{
The probability of exactly \( k \) successes.
}
\description{
Calculates the probability of exactly \( k \) successes in a binomial distribution.
}
\examples{
binomial_exactly(n = 15, p = 0.7, k = 10)
}
