% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_greater_than}
\alias{binomial_greater_than}
\title{Binomial Probability: Greater Than}
\usage{
binomial_greater_than(n, p, k)
}
\arguments{
\item{n}{The number of trials.}

\item{p}{The probability of success in a single trial.}

\item{k}{The threshold number of successes.}
}
\value{
The probability of more than \( k \) successes.
}
\description{
Calculates the probability of more than \( k \) successes in a binomial distribution.
}
\examples{
binomial_greater_than(n = 15, p = 0.7, k = 8)
}
