% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_strictly_between}
\alias{binomial_strictly_between}
\title{Binomial Probability: Between Two Values (Exclusive)}
\usage{
binomial_strictly_between(n, p, k1, k2)
}
\arguments{
\item{n}{The number of trials.}

\item{p}{The probability of success in a single trial.}

\item{k1}{The lower bound for successes (exclusive).}

\item{k2}{The upper bound for successes (exclusive).}
}
\value{
The probability of successes strictly between \( k1 \) and \( k2 \).
}
\description{
Calculates the probability of k1 < X < k2 successes in a binomial distribution.
}
\examples{
binomial_strictly_between(n = 20, p = 0.2, k1 = 7, k2 = 9)
}
