% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_E_XY}
\alias{calculate_E_XY}
\title{E(XY) Calculation}
\usage{
calculate_E_XY(f_xy, x_values, y_values)
}
\arguments{
\item{f_xy}{A function representing the joint probability distribution \( f(x, y) \).}

\item{x_values}{A numeric vector of possible X values.}

\item{y_values}{A numeric vector of possible Y values.}
}
\value{
The expected value \( E(XY) \) with a detailed explanation.
}
\description{
Calculates the expected value \( E(XY) \) for a given joint probability distribution.
}
\examples{
f_xy <- function(x, y) (x + 2 * y) / 42
x_values <- c(0, 1, 2, 3)
y_values <- c(0, 1, 2)
calculate_E_XY(f_xy, x_values, y_values)
}
