% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_lsrl}
\alias{calculate_lsrl}
\title{Compute the Least Squares Regression Line}
\usage{
calculate_lsrl(x, y)
}
\arguments{
\item{x}{A numeric vector of independent variable values.}

\item{y}{A numeric vector of dependent variable values.}
}
\value{
A list containing the LSRL equation, slope, intercept, and fitted values.
}
\description{
This function calculates the slope and intercept of the LSRL for given x and y data.
}
\examples{
calculate_lsrl(c(4, 4, 7, 12, 13, 20), c(23, 28, 30, 40, 28, 41))
}
