% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cdf_probability}
\alias{cdf_probability}
\title{Calculate Complementary Probability from CDF}
\usage{
cdf_probability(cdf, k)
}
\arguments{
\item{cdf}{A function representing the cumulative distribution function, defined for a range of \( x \).}

\item{k}{The threshold value for the probability calculation.}
}
\value{
A list containing the computed probability and a detailed explanation.
}
\description{
Computes the probability given a cumulative distribution function (CDF).
}
\examples{
# Define the CDF
cdf <- function(x) ifelse(x >= 0 & x <= 3, (x^3) / 27, ifelse(x > 3, 1, 0))
# Compute P(X >= 1)
cdf_probability(cdf, k = 1)
}
