% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci_diff_proportions}
\alias{ci_diff_proportions}
\title{Confidence Interval for the Difference in Proportions}
\usage{
ci_diff_proportions(
  successes1,
  sample_size1,
  successes2,
  sample_size2,
  confidence_level = 0.95
)
}
\arguments{
\item{successes1}{The number of successes in the first sample (integer).}

\item{sample_size1}{The size of the first sample (integer).}

\item{successes2}{The number of successes in the second sample (integer).}

\item{sample_size2}{The size of the second sample (integer).}

\item{confidence_level}{The confidence level (e.g., 0.95 for 95 percent confidence).}
}
\value{
A list with intermediate calculations and the confidence interval:
  \item{proportion1}{The first sample proportion (p1).}
  \item{proportion2}{The second sample proportion (p2).}
  \item{standard_error}{The standard error of the difference in proportions.}
  \item{z_value}{The critical z-value.}
  \item{margin_of_error}{The margin of error.}
  \item{lower_bound}{The lower bound of the confidence interval.}
  \item{upper_bound}{The upper bound of the confidence interval.}
}
\description{
Calculates the confidence interval for the difference between two population proportions and shows intermediate steps.
}
\examples{
ci_diff_proportions(30, 100, 20, 100, 0.95)

}
