% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci_median}
\alias{ci_median}
\title{Confidence Interval for Median}
\usage{
ci_median(data, confidence_level = 0.95, n_bootstrap = 1000)
}
\arguments{
\item{data}{A numeric vector of sample data.}

\item{confidence_level}{The confidence level (e.g., 0.95 for 95 percent confidence).}

\item{n_bootstrap}{The number of bootstrap samples (default is 1000).}
}
\value{
A list with intermediate calculations and the confidence interval:
  \item{bootstrap_medians}{A vector of bootstrapped medians.}
  \item{lower_bound}{The lower bound of the confidence interval.}
  \item{upper_bound}{The upper bound of the confidence interval.}
}
\description{
Calculates the confidence interval for the median using bootstrapping.
Intermediate steps are included in the output.
}
\examples{
ci_median(c(5, 10, 15, 20, 25), 0.95, 1000)

}
