% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci_to_new_ci}
\alias{ci_to_new_ci}
\title{Calculate the Confidence Interval for a New Confidence Level with Detailed Information}
\usage{
ci_to_new_ci(
  lower_original,
  upper_original,
  n,
  confidence_original,
  confidence_new
)
}
\arguments{
\item{lower_original}{The lower bound of the original confidence interval.}

\item{upper_original}{The upper bound of the original confidence interval.}

\item{n}{The sample size.}

\item{confidence_original}{The original confidence level (e.g., 0.90 for 90 percent).}

\item{confidence_new}{The new confidence level (e.g., 0.99 for 99 percent).}
}
\value{
A list containing the following:
  - `new_confidence_interval`: The new confidence interval as a vector.
  - `sample_mean`: The sample mean calculated from the original confidence interval.
  - `margin_error_original`: The margin of error for the original confidence interval.
  - `t_original`: The t-value for the original confidence level.
  - `t_new`: The t-value for the new confidence level.
  - `sample_standard_deviation`: The sample standard deviation.
  - `margin_error_new`: The margin of error for the new confidence interval.
}
\description{
This function calculates the confidence interval for a new confidence level based on a given original
confidence interval, sample size, and critical t-values for both the original and new confidence levels.
It uses the t-distribution to calculate the new margin of error and confidence interval.
}
\examples{
ci_to_new_ci(15.34, 36.66, 5, 0.90, 0.99)
}
