% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conditional_probability_events}
\alias{conditional_probability_events}
\title{Conditional Probability for Events}
\usage{
conditional_probability_events(p_A_and_B, p_B)
}
\arguments{
\item{p_A_and_B}{Probability of the intersection of events \( A \) and \( B \) (e.g., both high blood pressure and runner).}

\item{p_B}{Probability of event \( B \) (e.g., being a runner).}
}
\value{
A list containing the conditional probability \( P(A | B) \) and a detailed explanation.
}
\description{
Calculates the conditional probability of an event \( A \) given another event \( B \).
}
\examples{
conditional_probability_events(p_A_and_B = 0.1, p_B = 0.3)
}
