% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conditional_probability_intersection}
\alias{conditional_probability_intersection}
\title{Conditional Probability P(A|B) with Intersection}
\usage{
conditional_probability_intersection(p_intersection_ab, p_b)
}
\arguments{
\item{p_intersection_ab}{Probability of A intersection B.}

\item{p_b}{Probability of B.}
}
\value{
The conditional probability P(A|B).
}
\description{
Calculates the conditional probability P(A|B) using P(A intersection B) and P(B).
}
\examples{
conditional_probability_intersection(p_intersection_ab = 0.11, p_b = 0.46)
}
