% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conditional_probability_union}
\alias{conditional_probability_union}
\title{Conditional Probability P(A|B) with Union}
\usage{
conditional_probability_union(p_union_ab, p_a, p_b)
}
\arguments{
\item{p_union_ab}{Probability of A union B.}

\item{p_a}{Probability of A.}

\item{p_b}{Probability of B.}
}
\value{
The conditional probability P(A|B).
}
\description{
Calculates the conditional probability P(A|B) using P(A union B), P(A), and P(B).
}
\examples{
conditional_probability_union(p_union_ab = 0.62, p_a = 0.27, p_b = 0.46)
}
