% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_interval_true_mean}
\alias{confidence_interval_true_mean}
\title{Confidence Interval for True Mean}
\usage{
confidence_interval_true_mean(
  sample_mean,
  sample_variance,
  sample_size,
  confidence_level
)
}
\arguments{
\item{sample_mean}{The sample mean.}

\item{sample_variance}{The sample variance.}

\item{sample_size}{The size of the sample.}

\item{confidence_level}{The desired confidence level (e.g., 0.99 for 99 percent confidence).}
}
\value{
A list containing the lower and upper bounds of the confidence interval and a detailed explanation.
}
\description{
Calculates a confidence interval for the true mean given sample statistics.
}
\examples{
confidence_interval_true_mean(sample_mean = 11.3, sample_variance = 0.49,
                              sample_size = 12, confidence_level = 0.99)
}
