% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correlation_explanation}
\alias{correlation_explanation}
\title{Correlation Coefficient Explanation}
\usage{
correlation_explanation(r)
}
\arguments{
\item{r}{The correlation coefficient (numeric, between -1 and 1).}
}
\value{
An explanation of the relationship between two variables based on ( r).
}
\description{
Explains the meaning of the correlation coefficient ( r) when it is close to 1, -1, or 0.
}
\examples{
correlation_explanation(0.9)
correlation_explanation(0)
correlation_explanation(-0.8)
}
