% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{critical_t_value}
\alias{critical_t_value}
\title{Critical t* Value}
\usage{
critical_t_value(prob, df)
}
\arguments{
\item{prob}{The right-tail probability (a value between 0 and 1).}

\item{df}{The degrees of freedom (a positive integer).}
}
\value{
A list containing the critical t* value and an explanation of the calculation.
}
\description{
Calculates the critical t* value for a given right-tail probability and degrees of freedom.
}
\examples{
critical_t_value(prob = 0.10, df = 8)
}
