% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cumulative_probability_between}
\alias{cumulative_probability_between}
\title{cumulative Probability Between}
\usage{
cumulative_probability_between(x, p, thresholds)
}
\arguments{
\item{x}{A numeric vector of possible values of the random variable.}

\item{p}{A numeric vector of probabilities corresponding to the values in `x`.}

\item{thresholds}{A numeric vector specifying two thresholds (e.g., c(lower, upper)).}
}
\value{
A list containing the calculated probability and a detailed explanation.
}
\description{
Calculates the probability that a discrete random variable is between two thresholds (exclusive).
}
\examples{
cumulative_probability_between(x = c(0, 1, 2, 3), p = c(4, 5, 8, 3), thresholds = c(1, 3))
}
