% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discrete_between}
\alias{discrete_between}
\title{Discrete Probability: Between Two Values}
\usage{
discrete_between(prob_dist, x1, x2)
}
\arguments{
\item{prob_dist}{A named numeric vector where names represent the values of \( X \) and values represent \( P(X) \).}

\item{x1}{The lower bound for \( X \) (inclusive).}

\item{x2}{The upper bound for \( X \) (inclusive).}
}
\value{
The probability of x1 <= X <= x2.
}
\description{
Calculates the probability of x1 <= X <= x2 in a discrete probability distribution.
}
\examples{
prob_dist <- c("0" = 0.1, "1" = 0.2, "2" = 0.3, "3" = 0.4)
discrete_between(prob_dist, x1 = 1, x2 = 3)
}
