% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discrete_exactly}
\alias{discrete_exactly}
\title{Discrete Probability: Exactly}
\usage{
discrete_exactly(prob_dist, x)
}
\arguments{
\item{prob_dist}{A named numeric vector where names represent the values of \( X \) and values represent \( P(X) \).}

\item{x}{The value for which the probability is calculated.}
}
\value{
The probability of \( X = x \).
}
\description{
Calculates the probability of exactly \( x \) in a discrete probability distribution.
}
\examples{
prob_dist <- c("0" = 0.1, "1" = 0.2, "2" = 0.3, "3" = 0.4)
discrete_exactly(prob_dist, x = 2)
}
