% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_explanation}
\alias{error_explanation}
\title{Explanation of Type I and Type II Errors}
\usage{
error_explanation(null_hypothesis, alternative_hypothesis, error_type)
}
\arguments{
\item{null_hypothesis}{The null hypothesis.}

\item{alternative_hypothesis}{The alternative hypothesis.}

\item{error_type}{The type of error to explain: "Type I" or "Type II".}
}
\value{
A string explaining the selected type of error.
}
\description{
This function explains the concept of Type I and Type II errors in the context of hypothesis testing.
It provides an interpretation based on a given scenario, where the null hypothesis and alternative hypothesis are defined.
}
\examples{
error_explanation(
  "The team will not get the first down",
  "The team will get the first down",
  "Type I"
)
}
