% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expected_value_of_pdf}
\alias{expected_value_of_pdf}
\title{Compute Expected Value for a Continuous Random Variable}
\usage{
expected_value_of_pdf(pdf, lower, upper)
}
\arguments{
\item{pdf}{A function representing the probability density function.}

\item{lower}{The lower limit of the range of \( X \).}

\item{upper}{The upper limit of the range of \( X \).}
}
\value{
The expected value \( E[X] \).
}
\description{
Computes the expected value \( E[X] \) given a probability density function (PDF) and its range.
}
\examples{
pdf <- function(x) (x^2) / 72
expected_value_of_pdf(pdf, lower = 0, upper = 6)
}
