% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_c_for_probability_target}
\alias{find_c_for_probability_target}
\title{Find c such that P(Z < c) = target_probability}
\usage{
find_c_for_probability_target(target_probability)
}
\arguments{
\item{target_probability}{The target probability for which to find the value of c.}
}
\value{
The value of c such that P(Z < c) = target_probability.
}
\description{
This function finds the value of c such that P(Z < c) equals a given probability.
}
\examples{
find_c_for_probability_target(0.845)

}
