% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_c_greater}
\alias{find_c_greater}
\title{Find Z-score for P(Z > c)}
\usage{
find_c_greater(target_prob)
}
\arguments{
\item{target_prob}{Probability for P(Z > c), a value between 0 and 1 (exclusive).}
}
\value{
The Z-score c.
}
\description{
Calculates the Z-score c such that P(Z > c) = target_prob for the standard normal distribution.
}
\examples{
find_c_greater(0.65)
}
