% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_E_c}
\alias{find_E_c}
\title{Expected Value for Z-scores from CDF Probabilities}
\usage{
find_E_c(cumulative_probs, probabilities)
}
\arguments{
\item{cumulative_probs}{A vector of cumulative probabilities A.}

\item{probabilities}{A vector of probabilities for the values of A.}
}
\value{
The expected value E(c).
}
\description{
Calculates the expected value of Z-scores for given cumulative probabilities.
}
\examples{
find_E_c(cumulative_probs = c(0.67, 0.61), probabilities = c(0.75, 0.25))
}
