% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_x_between}
\alias{find_x_between}
\title{Find X Values Where P(lower < X < upper)}
\usage{
find_x_between(mean, sd, probability)
}
\arguments{
\item{mean}{The mean of the distribution (numeric).}

\item{sd}{The standard deviation of the distribution (numeric).}

\item{probability}{The target probability (numeric, must be between 0 and 1).}
}
\value{
A list containing the lower and upper `x` values corresponding to the specified probability.
}
\description{
Finds two `x` values such that the probability of a normally distributed random variable being between them equals the specified probability.
}
\examples{
find_x_between(mean = 100, sd = 15, probability = 0.95)
}
