% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_x_greater}
\alias{find_x_greater}
\title{Find X Value Where P(X > x)}
\usage{
find_x_greater(mean, sd, probability)
}
\arguments{
\item{mean}{The mean of the distribution (numeric).}

\item{sd}{The standard deviation of the distribution (numeric).}

\item{probability}{The target probability (numeric, must be between 0 and 1).}
}
\value{
The `x` value corresponding to the specified probability (numeric).
}
\description{
Finds the value of `x` such that the probability of a normally distributed random variable being greater than `x` equals the specified probability.
}
\examples{
find_x_greater(mean = 100, sd = 15, probability = 0.05)
}
