% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_z_between_exclusive}
\alias{find_z_between_exclusive}
\title{Find range for P(a < Z < b)}
\usage{
find_z_between_exclusive(target_prob)
}
\arguments{
\item{target_prob}{Probability for P(a < Z < b), a value between 0 and 1 (exclusive).}
}
\value{
A vector with lower and upper bounds, c(a, b).
}
\description{
Finds the Z-scores a and b such that P(a < Z < b) = target_prob for the standard normal distribution.
}
\examples{
find_z_between_exclusive(0.9)
}
