% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypergeo_between}
\alias{hypergeo_between}
\title{Hypergeometric Probability: Between Two Values}
\usage{
hypergeo_between(N, K, n, k1, k2)
}
\arguments{
\item{N}{The total population size.}

\item{K}{The total number of successes in the population.}

\item{n}{The number of draws.}

\item{k1}{The lower bound of successes (inclusive).}

\item{k2}{The upper bound of successes (inclusive).}
}
\value{
The probability of selecting between \( k1 \) and \( k2 \) successes.
}
\description{
Calculates the probability of selecting between \( k1 \) and \( k2 \) successes (inclusive) in a hypergeometric distribution.
}
\examples{
hypergeo_between(N = 13, K = 6, n = 6, k1 = 1, k2 = 3)
}
