% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypergeo_greater_than}
\alias{hypergeo_greater_than}
\title{Hypergeometric Probability: Greater Than}
\usage{
hypergeo_greater_than(N, K, n, k)
}
\arguments{
\item{N}{The total population size.}

\item{K}{The total number of successes in the population.}

\item{n}{The number of draws.}

\item{k}{The threshold number of successes.}
}
\value{
The probability of selecting more than \( k \) successes.
}
\description{
Calculates the probability of selecting more than \( k \) successes in a hypergeometric distribution.
}
\examples{
hypergeo_greater_than(N = 13, K = 6, n = 6, k = 2)
}
