% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypergeo_less_than}
\alias{hypergeo_less_than}
\title{Hypergeometric Probability: Less Than}
\usage{
hypergeo_less_than(N, K, n, k)
}
\arguments{
\item{N}{The total population size.}

\item{K}{The total number of successes in the population.}

\item{n}{The number of draws.}

\item{k}{The threshold number of successes (exclusive).}
}
\value{
The probability of selecting fewer than \( k \) successes.
}
\description{
Calculates the probability of selecting fewer than \( k \) successes in a hypergeometric distribution.
}
\examples{
hypergeo_less_than(N = 13, K = 6, n = 6, k = 3)
}
