% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{joint_between}
\alias{joint_between}
\title{Joint Probability: Between Two Values}
\usage{
joint_between(dist1, dist2, k1, k2)
}
\arguments{
\item{dist1}{A named numeric vector representing the probability distribution of the first variable.}

\item{dist2}{A named numeric vector representing the probability distribution of the second variable.}

\item{k1}{The lower bound for the sum (inclusive).}

\item{k2}{The upper bound for the sum (inclusive).}
}
\value{
The probability that the sum of the two variables is between \( k1 \) and \( k2 \).
}
\description{
Calculates the probability that the sum of two independent random variables is between \( k1 \) and \( k2 \).
}
\examples{
dist1 <- c("0" = 0.5, "1" = 0.1, "2" = 0.25, "3" = 0.15)
dist2 <- c("0" = 0.25, "1" = 0.45, "2" = 0.1, "3" = 0.2)
joint_between(dist1, dist2, k1 = 2, k2 = 4)
}
