% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{joint_less_than}
\alias{joint_less_than}
\title{Joint Probability: Less Than}
\usage{
joint_less_than(dist1, dist2, k)
}
\arguments{
\item{dist1}{A named numeric vector representing the probability distribution of the first variable.}

\item{dist2}{A named numeric vector representing the probability distribution of the second variable.}

\item{k}{The threshold for the sum of the two random variables.}
}
\value{
The probability that the sum of the two variables is less than \( k \).
}
\description{
Calculates the probability that the sum of two independent random variables is less than \( k \).
}
\examples{
dist1 <- c("0" = 0.5, "1" = 0.1, "2" = 0.25, "3" = 0.15)
dist2 <- c("0" = 0.25, "1" = 0.45, "2" = 0.1, "3" = 0.2)
joint_less_than(dist1, dist2, k = 2)
}
