% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linear_regression_summary}
\alias{linear_regression_summary}
\title{Perform Linear Regression Summary and Confidence Interval}
\usage{
linear_regression_summary(intercept, slope, slope_se, alpha = 0.05)
}
\arguments{
\item{intercept}{Estimate of the intercept from the regression output.}

\item{slope}{Estimate of the slope from the regression output.}

\item{slope_se}{Standard error of the slope from the regression output.}

\item{alpha}{Significance level (default is 0.05).}
}
\value{
A list containing the regression equation, slope confidence interval, and slope estimate.
}
\description{
This function provides the equation of the least squares regression line and the confidence interval for the slope.
}
\examples{
linear_regression_summary(intercept = -92.040,
slope = 0.710, slope_se = 0.041, alpha = 0.05)
}
