% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{median_and_shape}
\alias{median_and_shape}
\title{Median and Shape of Distribution}
\usage{
median_and_shape(data)
}
\arguments{
\item{data}{A numeric vector representing the dataset.}
}
\value{
A list containing the median and a description of the shape of the distribution.
}
\description{
Calculates the median and describes the shape of a distribution from a dataset.
}
\examples{
data <- c(2.3, 2.4, 2.6, 2.7, 2.7, 2.8, 2.8, 2.8, 2.9, 3.2, 3.2,
          3.3, 3.4, 3.6, 3.7, 3.8, 3.9, 4.1, 4.2, 4.2, 4.3, 4.4,
          5.3, 5.5, 5.5, 5.6)
median_and_shape(data)
}
