% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_between}
\alias{normal_between}
\title{Normal Distribution: Between Two Values}
\usage{
normal_between(mean, sd, lower, upper)
}
\arguments{
\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{lower}{The lower bound of the range.}

\item{upper}{The upper bound of the range.}
}
\value{
The probability that the random variable is between the given bounds.
}
\description{
Calculates the probability that a normal random variable is between two values.
}
\examples{
normal_between(mean = 0, sd = 1, lower = -1, upper = 1)
}
