% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_exact}
\alias{normal_exact}
\title{Normal Distribution: Exactly}
\usage{
normal_exact(mean, sd, value)
}
\arguments{
\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{value}{The value at which the probability is calculated.}
}
\value{
The probability that the random variable equals the given value (always zero for continuous variables).
}
\description{
Calculates the probability that a normal random variable equals a given value (theoretically zero for continuous distributions).
}
\examples{
normal_exact(mean = 0, sd = 1, value = 0)
}
