% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_greater}
\alias{normal_greater}
\title{Normal Distribution: Greater Than}
\usage{
normal_greater(mean, sd, value)
}
\arguments{
\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{value}{The value above which the probability is calculated.}
}
\value{
The probability that the random variable is greater than the given value.
}
\description{
Calculates the probability that a normal random variable is greater than a given value.
}
\examples{
normal_greater(mean = 0, sd = 1, value = 1.5)
}
