% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_less}
\alias{normal_less}
\title{Normal Distribution: Less Than}
\usage{
normal_less(mean, sd, value)
}
\arguments{
\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{value}{The value below which the probability is calculated.}
}
\value{
The probability that the random variable is less than the given value.
}
\description{
Calculates the probability that a normal random variable is less than a given value.
}
\examples{
normal_less(mean = 0, sd = 1, value = -1.5)
}
