% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_plot_standard_normal_distribution}
\alias{normal_plot_standard_normal_distribution}
\title{Plot the Distribution of Z}
\usage{
normal_plot_standard_normal_distribution(x_range = c(-4, 4))
}
\arguments{
\item{x_range}{The range of x values to plot the distribution.}
}
\value{
A plot of the standard normal distribution.
}
\description{
This function plots the standard normal distribution with a mean of 0 and a standard deviation of 1.
}
\examples{
normal_plot_standard_normal_distribution()

}
