% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_prob_Z_between_target_values}
\alias{normal_prob_Z_between_target_values}
\title{Find Probability P(lower < Z < upper)}
\usage{
normal_prob_Z_between_target_values(lower, upper)
}
\arguments{
\item{lower}{The lower bound for Z.}

\item{upper}{The upper bound for Z.}
}
\value{
The probability P(lower < Z < upper).
}
\description{
This function calculates the probability that the standard normal variable Z lies between two values, lower and upper.
}
\examples{
normal_prob_Z_between_target_values(-0.45, 1.96)

}
