% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normal_prob_Z_greater_than_target}
\alias{normal_prob_Z_greater_than_target}
\title{Find Probability P(Z > target)}
\usage{
normal_prob_Z_greater_than_target(target)
}
\arguments{
\item{target}{The value of Z for which to calculate the probability.}
}
\value{
The probability P(Z > target).
}
\description{
This function calculates the probability that the standard normal variable Z is greater than a given target.
}
\examples{
normal_prob_Z_greater_than_target(1.2)

}
