% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{one_sample_z_test_proportion}
\alias{one_sample_z_test_proportion}
\title{One-Sample z-Test for Proportion}
\usage{
one_sample_z_test_proportion(
  successes,
  sample_size,
  p,
  alternative = "two.sided"
)
}
\arguments{
\item{successes}{The number of successes in the sample.}

\item{sample_size}{The size of the sample.}

\item{p}{The hypothesized population proportion.}

\item{alternative}{The alternative hypothesis. Use "two.sided", "less", or "greater".}
}
\value{
A list containing the z-statistic and p-value.
}
\description{
Conducts a one-sample z-test for a population proportion.
}
\examples{
one_sample_z_test_proportion(successes = 50, sample_size = 100, p = 0.5, alternative = "two.sided")

}
