% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_conversion_check}
\alias{plot_conversion_check}
\title{Plot Conversion Checker}
\usage{
plot_conversion_check(plot1, plot2)
}
\arguments{
\item{plot1}{The first type of plot (e.g., "stem and leaf", "dot plot", "histogram").}

\item{plot2}{The second type of plot to evaluate (e.g., "histogram", "dot plot").}
}
\value{
A list containing the result (TRUE/FALSE) and a detailed explanation.
}
\description{
Determines if a second type of plot can always be generated from a first type of plot and provides a detailed explanation.
}
\examples{
plot_conversion_check("stem and leaf", "histogram")
plot_conversion_check("histogram", "dot plot")
}
