% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poisson_between}
\alias{poisson_between}
\title{Poisson Probability: Between Two Values}
\usage{
poisson_between(lambda, k1, k2)
}
\arguments{
\item{lambda}{The mean number of occurrences in the given time period.}

\item{k1}{The lower bound of events (inclusive).}

\item{k2}{The upper bound of events (inclusive).}
}
\value{
The probability of the random variable being between (k1) and (k2).
}
\description{
Calculates the probability of a Poisson random variable being between two values (inclusive).
}
\examples{
poisson_between(lambda = 3, k1 = 2, k2 = 4)
}
