% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poisson_between_3var_inclusive}
\alias{poisson_between_3var_inclusive}
\title{Poisson Probability Between Inclusive}
\usage{
poisson_between_3var_inclusive(rate, duration, a, b)
}
\arguments{
\item{rate}{The average rate of events per time unit.}

\item{duration}{The time period during which events are counted.}

\item{a}{The lower bound.}

\item{b}{The upper bound.}
}
\value{
The probability \( P(a X b) \).
}
\description{
Calculates the probability \( P(a X b) \) for a Poisson distribution.
}
\examples{
poisson_between_3var_inclusive(rate = 0.6, duration = 8, a = 3, b = 5)
}
