% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poisson_exact}
\alias{poisson_exact}
\title{Poisson Probability: Exactly}
\usage{
poisson_exact(lambda, k)
}
\arguments{
\item{lambda}{The mean number of occurrences in the given time period.}

\item{k}{The number of events for which the probability is calculated.}
}
\value{
The probability of exactly (k) events occurring.
}
\description{
Calculates the probability of exactly (k) events occurring in a Poisson distribution.
}
\examples{
poisson_exact(lambda = 3, k = 3)
}
