% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poisson_less}
\alias{poisson_less}
\title{Poisson Probability: Less Than}
\usage{
poisson_less(lambda, k)
}
\arguments{
\item{lambda}{The mean number of occurrences in the given time period.}

\item{k}{The threshold number of events (exclusive).}
}
\value{
The probability of fewer than (k) events occurring.
}
\description{
Calculates the probability of fewer than (k) events occurring in a Poisson distribution.
}
\examples{
poisson_less(lambda = 3, k = 3)
}
