% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poisson_less_3var}
\alias{poisson_less_3var}
\title{Poisson Probability Less Than}
\usage{
poisson_less_3var(rate, duration, k)
}
\arguments{
\item{rate}{The average rate of events per time unit.}

\item{duration}{The time period during which events are counted.}

\item{k}{The threshold value.}
}
\value{
The probability \( P(X < k) \).
}
\description{
Calculates the probability \( P(X < k) \) for a Poisson distribution.
}
\examples{
poisson_less_3var(rate = 0.6, duration = 8, k = 4)
}
