% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_sample_mean_less}
\alias{prob_sample_mean_less}
\title{Probability of Sample Mean Less Than a Threshold}
\usage{
prob_sample_mean_less(sample_mean, population_mean, population_sd, sample_size)
}
\arguments{
\item{sample_mean}{Threshold sample mean value to test.}

\item{population_mean}{Population mean (mu).}

\item{population_sd}{Population standard deviation (sigma).}

\item{sample_size}{Size of the sample (n).}
}
\value{
The probability that the sample mean is less than the threshold.
}
\description{
Calculates the probability that the sample mean is less than a threshold given population parameters and sample size.
}
\examples{
prob_sample_mean_less(sample_mean = 53,
population_mean = 47, population_sd = 10.2, sample_size = 25)
}
