% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{probability_sample_mean}
\alias{probability_sample_mean}
\title{Probability of Sample Mean}
\usage{
probability_sample_mean(
  sample_mean,
  population_mean,
  population_sd,
  sample_size
)
}
\arguments{
\item{sample_mean}{The sample mean to compare.}

\item{population_mean}{The population mean.}

\item{population_sd}{The population standard deviation.}

\item{sample_size}{The sample size.}
}
\value{
The probability ( P(Xbar < sample_mean) ).
}
\description{
Calculates the probability that the sample mean is less than a given value using the CLT.
}
\examples{
probability_sample_mean(12.01, 12.25, 2.14, 144)
}
