% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{probability_uniform_time_interval}
\alias{probability_uniform_time_interval}
\title{Probability of First Event in a Time Interval for Uniform Distribution}
\usage{
probability_uniform_time_interval(
  start_time,
  end_time,
  total_time = 1440,
  days = 1
)
}
\arguments{
\item{start_time}{Start time of the interval in "HH:MM" format.}

\item{end_time}{End time of the interval in "HH:MM" format.}

\item{total_time}{Total length of the uniform distribution period (e.g., 1440 minutes for a 24-hour day).}

\item{days}{Number of independent days.}
}
\value{
Probability that the event occurs within the interval for all days.
}
\description{
Calculates the probability that an event occurs within a given time interval for a uniform distribution, extended for multiple independent days.
}
\examples{
# Probability that the first drop of rain falls between 8:45 AM and 2:25 PM for two days
probability_uniform_time_interval("08:45", "14:25", total_time = 1440, days = 2)
}
