% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion}
\alias{proportion}
\title{Calculate Sample Proportion}
\usage{
proportion(successes, sample_size)
}
\arguments{
\item{successes}{The number of successes in the sample (integer).}

\item{sample_size}{The total size of the sample (integer).}
}
\value{
The sample proportion.
}
\description{
This function calculates the proportion of successes in a given sample.
}
\examples{
proportion(successes = 50, sample_size = 100)

}
