% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_confidence_interval}
\alias{proportion_confidence_interval}
\title{Confidence Interval for Proportion}
\usage{
proportion_confidence_interval(successes, sample_size, confidence_level)
}
\arguments{
\item{successes}{The number of successes in the sample (integer).}

\item{sample_size}{The total size of the sample (integer).}

\item{confidence_level}{The confidence level as a decimal (e.g., 0.95 for 95 percent confidence).}
}
\value{
A list containing the lower and upper bounds of the confidence interval.
}
\description{
Calculates the confidence interval for a sample proportion given the sample size and confidence level.
}
\examples{
proportion_confidence_interval(successes = 50, sample_size = 100, confidence_level = 0.95)

}
