% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_probability}
\alias{proportion_probability}
\title{Probability for Proportion}
\usage{
proportion_probability(p, sample_size, target, type = "less")
}
\arguments{
\item{p}{The population proportion.}

\item{sample_size}{The total size of the sample (integer).}

\item{target}{The target proportion for comparison.}

\item{type}{The type of probability to calculate. Use "less" for P(p_hat < target) or "greater" for P(p_hat > target).}
}
\value{
A list containing the z-score and the calculated probability.
}
\description{
Calculates the probability that the sample proportion is less than or greater than a given target value.
}
\examples{
proportion_probability(p = 0.5, sample_size = 100, target = 0.6, type = "less")

}
