% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_probability_between}
\alias{proportion_probability_between}
\title{Probability Between Two Proportions}
\usage{
proportion_probability_between(p, sample_size, lower, upper)
}
\arguments{
\item{p}{The population proportion.}

\item{sample_size}{The total size of the sample (integer).}

\item{lower}{The lower bound for the proportion.}

\item{upper}{The upper bound for the proportion.}
}
\value{
The probability that the sample proportion lies between the specified bounds.
}
\description{
Calculates the probability that the sample proportion lies between two specified values.
}
\examples{
proportion_probability_between(p = 0.5, sample_size = 100, lower = 0.4, upper = 0.6)

}
