% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_sd}
\alias{proportion_sd}
\title{Calculate Standard Error of the Proportion}
\usage{
proportion_sd(p_hat, sample_size)
}
\arguments{
\item{p_hat}{The sample proportion.}

\item{sample_size}{The total size of the sample (integer).}
}
\value{
The standard error of the proportion.
}
\description{
This function calculates the standard error of a proportion given the sample proportion and sample size.
}
\examples{
proportion_sd(p_hat = 0.5, sample_size = 100)

}
