% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_standard_error}
\alias{proportion_standard_error}
\title{Standard Error of Sample Proportion}
\usage{
proportion_standard_error(p_hat, n)
}
\arguments{
\item{p_hat}{The sample proportion (a value between 0 and 1).}

\item{n}{The sample size (a positive integer).}
}
\value{
A list containing the standard error and an explanation of the calculation.
}
\description{
Calculates the standard error of a sample proportion based on the proportion and sample size.
}
\examples{
proportion_standard_error(p_hat = 0.25, n = 200)
}
