% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_z_score}
\alias{proportion_z_score}
\title{Z-Score for Proportion}
\usage{
proportion_z_score(p_hat, p, sample_size)
}
\arguments{
\item{p_hat}{The sample proportion.}

\item{p}{The population proportion.}

\item{sample_size}{The total size of the sample (integer).}
}
\value{
The z-score for the sample proportion.
}
\description{
Calculates the z-score for a given sample proportion.
}
\examples{
proportion_z_score(p_hat = 0.6, p = 0.5, sample_size = 100)

}
