% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{regression_analysis_3_vector}
\alias{regression_analysis_3_vector}
\title{Regression Analysis for GPA Prediction}
\usage{
regression_analysis_3_vector(IT, RR, GPA)
}
\arguments{
\item{IT}{A numeric vector of intelligence test scores.}

\item{RR}{A numeric vector of reading rates.}

\item{GPA}{A numeric vector of grade-point averages.}
}
\value{
A detailed report for both regression models and a comparison of their fit based on \( R^2 \) values.
}
\description{
Calculates and compares lines of best fit for predicting GPA based on reading rates (RR) and intelligence test (IT) scores.
}
\examples{
IT <- c(155, 206, 156, 172, 209, 191, 163, 207)
RR <- c(43, 33, 27, 35, 25, 36, 36, 26)
GPA <- c(3.0, 2.0, 1.7, 2.5, 2.0, 2.7, 2.6, 1.8)
regression_analysis_3_vector(IT, RR, GPA)
}
