% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean}
\alias{sample_mean}
\title{Calculate Sample Mean}
\usage{
sample_mean(data)
}
\arguments{
\item{data}{A numeric vector of sample data.}
}
\value{
The sample mean.
}
\description{
This function calculates the mean of a given sample.
}
\examples{
sample_mean(c(5, 10, 15))

}
