% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_confidence_interval}
\alias{sample_mean_confidence_interval}
\title{Confidence Interval for Sample Mean}
\usage{
sample_mean_confidence_interval(mean, sd, n, confidence_level)
}
\arguments{
\item{mean}{The sample mean.}

\item{sd}{The population standard deviation.}

\item{n}{The sample size.}

\item{confidence_level}{The confidence level as a decimal (e.g., 0.95 for 95 percent confidence).}
}
\value{
A list containing the lower and upper bounds of the confidence interval.
}
\description{
Calculates the confidence interval for the sample mean given the sample mean, population standard deviation, sample size, and confidence level.
}
\examples{
sample_mean_confidence_interval(mean = 20, sd = 5, n = 30, confidence_level = 0.95)

}
