% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_probability}
\alias{sample_mean_probability}
\title{Probability for Sample Mean}
\usage{
sample_mean_probability(mu, sd, n, target, type = "less")
}
\arguments{
\item{mu}{The population mean.}

\item{sd}{The population standard deviation.}

\item{n}{The sample size.}

\item{target}{The target value for the sample mean.}

\item{type}{The type of probability to calculate. Use "less" for P(X < target) or "greater" for P(X > target).}
}
\value{
A list containing the z-score and the calculated probability.
}
\description{
Calculates the probability that the sample mean is less than or greater than a given target value.
}
\examples{
sample_mean_probability(mu = 100, sd = 15, n = 30, target = 105, type = "less")

}
