% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_probability_between}
\alias{sample_mean_probability_between}
\title{Probability Between Two Values for Sample Mean}
\usage{
sample_mean_probability_between(mu, sd, n, lower, upper)
}
\arguments{
\item{mu}{The population mean.}

\item{sd}{The population standard deviation.}

\item{n}{The sample size.}

\item{lower}{The lower bound for the sample mean.}

\item{upper}{The upper bound for the sample mean.}
}
\value{
The probability that the sample mean lies between the specified bounds.
}
\description{
Calculates the probability that the sample mean lies between two specified values.
}
\examples{
sample_mean_probability_between(mu = 100, sd = 15, n = 30, lower = 95, upper = 105)

}
