% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_required_size}
\alias{sample_mean_required_size}
\title{Required Sample Size for Sample Mean}
\usage{
sample_mean_required_size(sd, margin_of_error, confidence_level)
}
\arguments{
\item{sd}{The standard deviation of the population.}

\item{margin_of_error}{The desired margin of error.}

\item{confidence_level}{The confidence level as a decimal (e.g., 0.95 for 95 percent confidence).}
}
\value{
The required sample size as an integer.
}
\description{
Calculates the required sample size to estimate the sample mean with a specified margin of error and confidence level.
}
\examples{
sample_mean_required_size(sd = 10, margin_of_error = 2, confidence_level = 0.95)

}
