% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_sd}
\alias{sample_mean_sd}
\title{Calculate Standard Error of the Sample Mean}
\usage{
sample_mean_sd(sd, n)
}
\arguments{
\item{sd}{The standard deviation of the population.}

\item{n}{The sample size.}
}
\value{
The standard error of the sample mean.
}
\description{
This function calculates the standard error of the sample mean given the population standard deviation and sample size.
}
\examples{
sample_mean_sd(sd = 10, n = 25)

}
