% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean_z_score}
\alias{sample_mean_z_score}
\title{Z-Score for Sample Mean}
\usage{
sample_mean_z_score(mean, mu, sd, n)
}
\arguments{
\item{mean}{The sample mean.}

\item{mu}{The population mean.}

\item{sd}{The population standard deviation.}

\item{n}{The sample size.}
}
\value{
The z-score for the sample mean.
}
\description{
Calculates the z-score for a given sample mean.
}
\examples{
sample_mean_z_score(mean = 105, mu = 100, sd = 15, n = 30)

}
