% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_size_by_width_cl}
\alias{sample_size_by_width_cl}
\title{Calculate the Margin of Error for a Confidence Interval Without a Sample Size}
\usage{
sample_size_by_width_cl(width, confidence_level, p_estimate = 0.5)
}
\arguments{
\item{width}{The desired total width of the confidence interval (numeric).}

\item{confidence_level}{The desired confidence level (numeric, between 0 and 1).}

\item{p_estimate}{The estimated proportion (default is 0.5 for maximum uncertainty).}
}
\value{
The calculated margin of error (numeric).
}
\description{
This function calculates the margin of error for a population proportion given
the confidence level, desired confidence interval width, and estimated proportion.
}
\examples{
sample_size_by_width_cl(width = 0.05, confidence_level = 0.95)
}
