% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampling_distribution_mean}
\alias{sampling_distribution_mean}
\title{Calculate Mean of a Sampling Distribution}
\usage{
sampling_distribution_mean(x, p)
}
\arguments{
\item{x}{A numeric vector representing the values of the random variable.}

\item{p}{A numeric vector representing the probabilities corresponding to each value of `x`. One probability can be `NA` if missing.}
}
\value{
A list containing the mean of the sampling distribution and a detailed explanation.
}
\description{
Computes the mean (expected value) of a sampling distribution and validates the probabilities.
}
\examples{
x <- c(-16, -9, -5, 8, 17)
p <- c(1/100, 3/50, 7/100, 7/100, NA)
sampling_distribution_mean(x, p)
}
