% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampling_distribution_variance}
\alias{sampling_distribution_variance}
\title{Calculate Variance of a Sampling Distribution}
\usage{
sampling_distribution_variance(x, p)
}
\arguments{
\item{x}{A numeric vector representing the values of the random variable.}

\item{p}{A numeric vector representing the probabilities corresponding to each value of `x`.}
}
\value{
A list containing the variance, mean, and a detailed explanation.
}
\description{
Computes the variance of a sampling distribution given its values and probabilities.
}
\examples{
x <- c(0, 10, 50, 100)
p <- c(1/14, 1/7, 2/7, 1/2)
sampling_distribution_variance(x, p)
}
