% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solve_for_a}
\alias{solve_for_a}
\title{Solve for Valid Parameter of a PDF}
\usage{
solve_for_a(f, lower = 0)
}
\arguments{
\item{f}{A function representing the pdf \( f(x) \).}

\item{lower}{The lower bound of the range for \( x \) (default = 0).}
}
\value{
The value of \( a \) that makes the pdf valid and an explanation.
}
\description{
Solves for the parameter \( a \) such that the given pdf \( f(x) \) is valid.
}
\examples{
pdf <- function(x) (1/2) * x^2
solve_for_a(pdf, lower = 0)
}
