% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standard_error_difference_means}
\alias{standard_error_difference_means}
\title{Standard Error for the Difference in Means}
\usage{
standard_error_difference_means(sd1, n1, sd2, n2)
}
\arguments{
\item{sd1}{Standard deviation of the first group.}

\item{n1}{Sample size of the first group.}

\item{sd2}{Standard deviation of the second group.}

\item{n2}{Sample size of the second group.}
}
\value{
A list containing the standard error and detailed explanation.
}
\description{
Calculates the standard error for the difference in two means.
}
\examples{
standard_error_difference_means(sd1 = 15, n1 = 30, sd2 = 20, n2 = 40)
}
