% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standard_error_difference_proportions}
\alias{standard_error_difference_proportions}
\title{Standard Error for the Difference in Proportions}
\usage{
standard_error_difference_proportions(p1, n1, p2, n2)
}
\arguments{
\item{p1}{Proportion of the first group.}

\item{n1}{Sample size of the first group.}

\item{p2}{Proportion of the second group.}

\item{n2}{Sample size of the second group.}
}
\value{
A list containing the standard error and detailed explanation.
}
\description{
Calculates the standard error for the difference in two proportions.
}
\examples{
standard_error_difference_proportions(p1 = 0.6, n1 = 50, p2 = 0.5, n2 = 60)
}
