% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standard_error_mean}
\alias{standard_error_mean}
\title{Standard Error for a Single Mean}
\usage{
standard_error_mean(sd, n)
}
\arguments{
\item{sd}{Standard deviation of the sample.}

\item{n}{Sample size.}
}
\value{
A list containing the standard error and detailed explanation.
}
\description{
Calculates the standard error for a single mean.
}
\examples{
standard_error_mean(sd = 10, n = 50)
}
