% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standard_error_proportion}
\alias{standard_error_proportion}
\title{Standard Error for a Single Proportion}
\usage{
standard_error_proportion(p, n)
}
\arguments{
\item{p}{Proportion.}

\item{n}{Sample size.}
}
\value{
A list containing the standard error and detailed explanation.
}
\description{
Calculates the standard error for a single proportion.
}
\examples{
standard_error_proportion(p = 0.5, n = 100)
}
