% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stats3339_query}
\alias{stats3339_query}
\title{API Prompt Function}
\usage{
stats3339_query(api_url, prompt_text)
}
\arguments{
\item{api_url}{The URL of the API to send the request to.}

\item{prompt_text}{The text prompt to send to the API.}
}
\value{
A character string containing the API response.
}
\description{
Sends a prompt to a specified API and retrieves the response.
}
