% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stats3339_query_apollo}
\alias{stats3339_query_apollo}
\title{API Prompt Function}
\usage{
stats3339_query_apollo(prompt_text)
}
\arguments{
\item{prompt_text}{The text prompt to send to the API.}
}
\value{
A character string containing the API response.
}
\description{
Sends a prompt to a specified API and retrieves the response.
}
