% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{two_sample_z_test_proportions}
\alias{two_sample_z_test_proportions}
\title{Two-Sample z-Test for Proportions}
\usage{
two_sample_z_test_proportions(
  successes1,
  sample_size1,
  successes2,
  sample_size2,
  alternative = "two.sided"
)
}
\arguments{
\item{successes1}{The number of successes in the first sample.}

\item{sample_size1}{The size of the first sample.}

\item{successes2}{The number of successes in the second sample.}

\item{sample_size2}{The size of the second sample.}

\item{alternative}{The alternative hypothesis. Use "two.sided", "less", or "greater".}
}
\value{
A list containing the z-statistic and p-value.
}
\description{
Conducts a two-sample z-test for the difference in population proportions.
}
\examples{
two_sample_z_test_proportions(successes1 = 50,
sample_size1 = 100,
successes2 = 60,
sample_size2 = 120,
alternative = "two.sided")

}
