% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniform_prob_time_greater_than}
\alias{uniform_prob_time_greater_than}
\title{Probability Greater Than a Value in a Uniform Distribution}
\usage{
uniform_prob_time_greater_than(a, b, x)
}
\arguments{
\item{a}{The lower bound of the uniform distribution.}

\item{b}{The upper bound of the uniform distribution.}

\item{x}{The value to compare.}
}
\value{
The probability that the variable is greater than `x`.
}
\description{
Calculates the probability of a uniform random variable being greater than a given value.
}
\examples{
uniform_prob_time_greater_than(0, 24, 14.4167) # Greater than 2:25 PM
}
