% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniform_prob_time_less_than}
\alias{uniform_prob_time_less_than}
\title{Probability Less Than a Value in a Uniform Distribution}
\usage{
uniform_prob_time_less_than(a, b, x)
}
\arguments{
\item{a}{The lower bound of the uniform distribution.}

\item{b}{The upper bound of the uniform distribution.}

\item{x}{The value to compare.}
}
\value{
The probability that the variable is less than `x`.
}
\description{
Calculates the probability of a uniform random variable being less than a given value.
}
\examples{
uniform_prob_time_less_than(0, 24, 8.75) # Less than 8:45 AM
}
